Do not break API and use internal utils instead
