+++ src/commands/synthetics/reporters/default.ts
@@ -19,12 +19,12 @@ import {
   UserConfigOverride,
   Batch,
 } from '../interfaces'
+import {hasResult} from '../utils/internal'
 import {
   getBatchUrl,
   getResultDuration,
   getResultOutcome,
   getResultUrl,
-  hasResult,
   isDeviceIdSet,
   isResultSkippedBySelectiveRerun,
   PASSED_RESULT_OUTCOMES,
