+++ src/commands/synthetics/utils/internal.ts
@@ -1,6 +1,14 @@
-import {ExecutionRule, Test, UserConfigOverride} from '../interfaces'
+import {
+  BaseResult,
+  ExecutionRule,
+  Result,
+  ResultInBatch,
+  ResultInBatchSkippedBySelectiveRerun,
+  Test,
+  UserConfigOverride,
+} from '../interfaces'
 
-import {getStrictestExecutionRule} from './public'
+import {getStrictestExecutionRule, isResultSkippedBySelectiveRerun} from './public'
 
 export const getOverriddenExecutionRule = (
   test?: Test,
@@ -9,4 +17,22 @@ export const getOverriddenExecutionRule = (
   if (configOverride?.executionRule) {
     return getStrictestExecutionRule(configOverride.executionRule, test?.options?.ci?.executionRule)
   }
+}
+
+export const hasResult = (result: Result): result is BaseResult => {
+  return !isResultSkippedBySelectiveRerun(result)
+}
+
+export const isResultInBatchSkippedBySelectiveRerun = (
+  result: ResultInBatch
+): result is ResultInBatchSkippedBySelectiveRerun => {
+  return result.selective_rerun?.decision === 'skip'
+}
+
+export const getResultIdOrLinkedResultId = (result: ResultInBatch): string => {
+  if (isResultInBatchSkippedBySelectiveRerun(result)) {
+    return result.selective_rerun.linked_result_id
+  }
+
+  return result.result_id
 }
